Automatic Travis documentation build
syn: simplify identifier validity checks

This new implementation brings a performance win of around 24-32% in
debug mode and is as fast as the old one in release mode.

As a result, now there is one fewer case in `build.rs` which should
reduce build time, albeit by a very little amount.